==> notebooks/01_load_validate_raw_data.ipynb <==
from pathlib import Path
import requests

def download_file_of_raw_data(year: int, month: int) -> Path:
    url = f'https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet'
    res = requests.get(url)

    if res.status_code == 200:
        path = f'../data/raw/rides_{year}-{month:02d}.parquet'
        open(path, "wb").write(res.content)
        return path
    else: 
        raise Exception(f'{url} is not available.') 

download_file_of_raw_data(2023, month=1)

import pandas as pd

rides = pd.read_parquet('../data/raw/rides_2023-01.parquet')

#rides.head(20)

#getting the two columns of time and place

rides = rides[['tpep_pickup_datetime', 'PULocationID']]

rides.rename(columns={
    'tpep_pickup_datetime': 'pickup_datetime',
    'PULocationID': 'pickup_location_id'
}, inplace = True)

#rides

#describes a summary of statistics regarding this column

rides['pickup_datetime'].describe()

#remove pickup dates that are not from specific year or month 
#filter data

rides = rides[rides.pickup_datetime >= '2023-01-01']
rides = rides[rides.pickup_datetime < '2023-02-01']

#rerun the program to check if the problem is fixed
rides['pickup_datetime'].describe()

#save the fixed raw data into new file

rides.to_parquet('../data/transformed/rides_2023-01.parquet')
