==> notebooks/05_vizualize_training_data.ipynb <==
%reload_ext autoreload
%autoreload 2

import os, sys

sys.path.append('c:\\Projects\\taxi_predicts\\src\\')
 

import pandas as pd
from paths import TRANSFORMED_DATA_DIR

features_and_targets = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
features_and_targets

features = features_and_targets.drop(columns=['target_rides_next_hour'])
targets = features_and_targets['target_rides_next_hour']

from plot import plot_one_sample

plot_one_sample(
    0,
    features,
    targets
)
