==> notebooks/06_baseline_model.ipynb <==
import os, sys

sys.path.append('c:\\Projects\\taxi_predicts\\src\\')

import pandas as pd
from paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

from datetime import datetime
from data_split import train_test_split

x_train, y_train, x_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2023, 6, 1, 0, 0, 0),
    target_column_name="target_rides_next_hour"
)

print(f'{x_train.shape=}')
print(f'{y_train.shape=}')
print(f'{x_test.shape=}')
print(f'{y_test.shape=}')

import numpy as np
import pandas as pd

#we are creating a simple model
#without any integrated functions
class BaselineModelPreviousHour:
    def fit(self, x_train: pd.DataFrame, y_train: pd.DataFrame):
        pass

    def predict(self, x_test: pd.DataFrame) -> np.array:
        return x_test[f'rides_previous_1_hour']
        

model = BaselineModelPreviousHour() #create object
predictions = model.predict(x_test) #predict the test features
predictions #this is the vector of predictions

#we need to estimate how good predictions are

#mean absolute error

from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

#still we are not gonna use machine learning,
#we are going to try to find a smarter rule base

import pandas as pd
import numpy as np

#we are creating a simple model
#without any integrated functions
class BaselineModelPreviousWeek:
    def fit(self, x_train: pd.DataFrame, y_train: pd.DataFrame):
        pass

    def predict(self, x_test: pd.DataFrame) -> np.array:
        return x_test[f'rides_previous_{7*24}_hour']
        

model = BaselineModelPreviousWeek() #create object
predictions = model.predict(x_test) #predict the test features
predictions #this is the vector of predictions

#the smaller the mae is the closer the predictions are from the actual values
#difference between our guesses and the actual answers

test_mae = mean_absolute_error(y_test, predictions) #computing errors
print(f'{test_mae=:.4f}')

import pandas as pd
import numpy as np

#we are creating a simple model with the last 7,14, 21,28 days
class BaselineModelLast4Weeks:
    def fit(self, x_train: pd.DataFrame, y_train: pd.DataFrame):
        pass

    def predict(self, x_test: pd.DataFrame) -> np.array:
        return 0.25*(
            #every week returns the avereage
            x_test[f'rides_previous_{7*24}_hour'] + \
            x_test[f'rides_previous_{7*2*24}_hour'] + \
            x_test[f'rides_previous_{7*3*24}_hour'] + \
            x_test[f'rides_previous_{7*4*24}_hour']
        )
        

model = BaselineModelLast4Weeks() #create object
predictions = model.predict(x_test) #predict the test features
predictions #this is the vector of predictions

from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')
