==> notebooks/09_lightbm_with_feature_engineering.ipynb <==
import os, sys
os.getcwd() #current working directory, path
sys.path.append('c:\\Projects\\taxi_predicts\\src')


import pandas as pd
from paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

from datetime import datetime
from data_split import train_test_split

x_train, y_train, x_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2023, 6, 1, 0, 0, 0),
    target_column_name="target_rides_next_hour"
)

print(f'{x_train.shape=}')
print(f'{y_train.shape=}')
print(f'{x_test.shape=}')
print(f'{y_test.shape=}')

def average_rides_last_4_weeks(x: pd.DataFrame) -> pd.DataFrame:
    #Adds one column with the average rides from 7, 14, 21, 28 days ago
    #regular python function
    
    x['average_rides_last_4_weeks'] = 0.25 * (
        x[f'rides_previous_{7*24}_hour'] + \
        x[f'rides_previous_{2*7*24}_hour'] + \
        x[f'rides_previous_{3*7*24}_hour'] + \
        x[f'rides_previous_{4*7*24}_hour']
    )

    return x

#we have to wrap our function to do the data transformation in scikitlearn
from sklearn.preprocessing import FunctionTransformer

add_feature_average_rides_last_4_weeks = FunctionTransformer(
    average_rides_last_4_weeks, validate=False)

add_feature_average_rides_last_4_weeks.fit_transform(x_train)

from sklearn.base import BaseEstimator, TransformerMixin
#lightbm is model that only accept only numerical features
#so we remove datetime

class TemporalFeaturesEngineer(BaseEstimator, TransformerMixin):

    def fit(self, x, y=None):
        return self
    
    def transform(self, x, y=None):
        x_ = x.copy()

        #generate numeric columns from datetime
        x_["hour"] = x_['pickup_hour'].dt.hour
        x_['day_of_week'] = x_['pickup_hour'].dt.dayofweek

        return x_.drop(columns=['pickup_hour'])

add_temporal_features = TemporalFeaturesEngineer()
add_temporal_features.fit_transform(x_train)

#now we need to add everything into pipeline

import lightgbm as lgb


from sklearn.pipeline import make_pipeline
#the training
pipeline = make_pipeline(
    add_feature_average_rides_last_4_weeks,
    add_temporal_features,
    lgb.LGBMRegressor()
)

pipeline.fit(x_train, y_train)

predictions = pipeline.predict(x_test)

from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

# https://towardsdatascience.com/how-to-quickly-design-advanced-sklearn-pipelines-3cc97b59ce16
