==> notebooks/13_model_training_pipeline.ipynb <==
import os, sys

sys.path.append('c:\\Projects\\taxi_predicts\\src\\')

%reload_ext autoreload
%autoreload 2

import config as config

import hopsworks
#now we connect to the feture_store and feature group to store the data

#connect to the project
project = hopsworks.login(
    project=config.HOPSWORKS_PROJECT_NAME,
    api_key_value=config.HOPSWORKS_API_KEY
)

#connect to the feature store
feature_store = project.get_feature_store()

#connect to the feature group
feature_group = feature_store.get_feature_group(
    name=config.FEATURE_GROUP_NAME,
    version=config.FEATURE_GROUP_VERSION,
)



#feature view, we define how we want to fetch data from feature groups and merge it to get the final data set

try:
    feature_store.create_feature_view(
        name=config.FEATURE_VIEW_NAME,
        version=config.FEATURE_VIEW_VERSION,
        query=feature_group.select_all()
    )
except:
    print("Feature view already existed. Skip creation.")


feature_view = feature_store.get_feature_view(
    name=config.FEATURE_VIEW_NAME,
    version=config.FEATURE_VIEW_VERSION
)

#we generate the training data with hopsowork method trainind_data

ts_data, _ = feature_view.training_data(
    description = "Time-series hourly taxi rides"
)

ts_data.sort_values(by=['pickup_location_id', 'pickup_hour'], inplace=True)
ts_data

from data import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*28,
    step_size=23
)

features_and_target = features.copy()
features_and_target['target_rides_next_hour'] = targets

print(f'{features_and_target.shape=}')

#now we split it into training and test data

from datetime import date, timedelta
from pytz import timezone
import pandas as pd
from data_split import train_test_split

#we are taking the data from today until one month ago and split it
cutoff_date = pd.to_datetime(date.today() -timedelta(days=28*1))

print(f'{cutoff_date=}')

x_train, y_train, x_test, y_test = train_test_split(
    features_and_target,
    cutoff_date,
    target_column_name='target_rides_next_hour'
)

print(f'{x_train.shape=}')
print(f'{x_test.shape=}')
print(f'{y_train.shape=}')
print(f'{y_test.shape=}')

#now we train the model with lgbmodel with feature engineering and hyperparameter tuning

import numpy as np
from sklearn.model_selection import KFold, TimeSeriesSplit
from sklearn.pipeline import make_pipeline
from sklearn.metrics import mean_absolute_error
import optuna #we use optuna for hyperparameters

from model import get_pipeline

def objective(trial: optuna.trial.Trial) -> float:
    """
    Given a set of hyper-parameters, it trains a model and computes an average
    validation error based on a TimeSeriesSplit
    """
    # pick hyper-parameters
    #we use optuna to find the best hyperparameters
    hyperparams = {
        "metric": 'mae',
        "verbose": -1,
        "num_leaves": trial.suggest_int("num_leaves", 2, 256),
        "feature_fraction": trial.suggest_float("feature_fraction", 0.2, 1.0),
        "bagging_fraction": trial.suggest_float("bagging_fraction", 0.2, 1.0),
        "min_child_samples": trial.suggest_int("min_child_samples", 3, 100),   
    }
    
    #we create a traning validation splits
    tss = TimeSeriesSplit(n_splits=4)
    scores = []
    
    # sort X_train by `pikup_hour` inplace
    # so the TimeSeriesSplit will split the data in a consistent way
    x_train.sort_values('pickup_hour', inplace=True)

    for train_index, val_index in tss.split(x_train):

        # split data for training and validation
        x_train_, x_val = x_train.iloc[train_index, :], x_train.iloc[val_index,:]
        y_train_, y_val_ = y_train.iloc[train_index], y_train.iloc[val_index]
        
        # train the model
        pipeline = get_pipeline(**hyperparams)
        pipeline.fit(x_train_, y_train_)
        
        # evaluate the model
        y_pred = pipeline.predict(x_val)
        mae = mean_absolute_error(y_val_, y_pred)

        scores.append(mae)
   
    # Return the mean score
    return np.array(scores).mean()

#study

study = optuna.create_study(direction="minimize")
study.optimize(objective, n_trials=1)

best_params = study.best_trial.params
print(f'{best_params=}')

pipeline = get_pipeline(**best_params)
pipeline.fit(x_train, y_train)

predictions = pipeline.predict(x_test)
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

#now we need to push the model into hopsworks registry model ad not localy
%reload_ext autoreload
%autoreload 2
import joblib
from paths import MODELS_DIR

joblib.dump(pipeline, MODELS_DIR / 'model.pkl')

#the modelSchema is method for hopsworks
from hsml.schema import Schema
from hsml.model_schema import ModelSchema

input_schema = Schema(x_train)
output_schema = Schema(y_train)
model_schema = ModelSchema(input_schema=input_schema, output_schema=output_schema)

#and now we push it to the model registry
#didnt work from video

model_registry = project.get_model_registry()

model = model_registry.sklearn.create_model(
    name='taxi_demand_predictor_next_hour',
    metrics={"test_mae": test_mae},
    description="LightGBM regressor with a bit of hyper-parameter tuning",
    input_example=x_train.sample(),
    model_schema=model_schema
)

model.save('../models/model.pkl')
# model.save(MODELS_DIR / 'model.pkl') #not working error
