==> notebooks/14_inference_pipeline.ipynb <==
%load_ext autoreload
%autoreload 2

import os, sys

sys.path.append('c:\\Projects\\taxi_predicts\\src\\')

from datetime import datetime, timedelta
import pandas as pd

current_date = pd.to_datetime(datetime.utcnow()).floor('H') # - timedelta(hours=1)
print(f'{current_date=}')
# current_date = pd.Timestamp('2023-02-28 09:00:00')

from inference import load_batch_of_features_from_store

#loading batch of features from store
features = load_batch_of_features_from_store(current_date)

from inference import (
    load_model_from_registry,
    get_model_predictions
)

model = load_model_from_registry()
predictions = get_model_predictions(model, features)

predictions['pickup_hour'] = current_date
predictions

import hopsworks
import config as config

project = hopsworks.login(
    project=config.HOPSWORKS_PROJECT_NAME,
    api_key_value=config.HOPSWORKS_API_KEY
)

feature_store = project.get_feature_store()


# connect to the feature group
feature_group = feature_store.get_or_create_feature_group(
    name='model_predictions_feature_group',
    version=1,
    description="Predictions generate by our production model",
    primary_key = ['pickup_hour', 'pickup_location_id'],
    event_time='pickup_hour',
)

predictions.head(10)

#we put the predictions into feature store so they can be later used 
feature_group.insert(predictions, write_options={"wait_for_job": False})

#we need to run this everytime the feature pipeline completes so we create one more action
